+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
 ## v0.18.0
+
 - fields/properties can be added posterior to instantiation of a Thing object
 - semantically identical "things" (same URI) can be converted into each other
 - fixed coverage issues and documentation
+- add convenience function `dump_model_ttl()` to dump a Thing object to turtle format
 
 ## v0.17.0